Fix testLuaDestructors test case on architecures other than amd64/i386.
The test suite tests if a destructor is called in a lambda called by
lua. However, the test object never runs out of scope, as the called
luaL_error() never returns. (see lua documentation)
The patch puts the test object in a scope, so that the compiler is allowed
to clean up the object.

(This unit tests interestingly fails only on most of the architectures,
but not on amd64/i386)